generation dashboard - initial
