+++ .gitignore
@@ -1,3 +1,4 @@
 .env
 venv*
 *__pycache__*
+tmp
