+++ dashboards/utils.py
@@ -1,0 +1,90 @@
+from glob import glob
+
+import yaml
+import torch
+import streamlit as st
+from datasets import Dataset, load_dataset
+
+
+def select_model_and_device():
+    with st.sidebar:
+        st.header("Model Configuration")
+        devices = [f"cuda:{it}" for it in range(torch.cuda.device_count())] + ["cpu"]
+        device = st.selectbox("Select Device", options=devices, help="Choose the device to run the model on")
+        checkpoint_path = st.selectbox(
+            "Select Checkpoint",
+            options=glob("checkpoints/*.pt"),
+            help="Choose the model checkpoint to use",
+        )
+
+    return device, checkpoint_path
+
+
+def select_part_dataset(midi_dataset: Dataset) -> Dataset:
+    """
+    Allows the user to select a part of the dataset based on composer and title.
+
+    Parameters:
+        midi_dataset (Dataset): The MIDI dataset to select from.
+
+    Returns:
+        Dataset: The selected part of the dataset.
+    """
+    source_df = midi_dataset.to_pandas()
+    source_df["source"] = source_df["source"].map(lambda source: yaml.safe_load(source))
+    source_df["composer"] = [source["composer"] for source in source_df.source]
+    source_df["title"] = [source["title"] for source in source_df.source]
+
+    composers = source_df.composer.unique()
+    selected_composer = st.selectbox(
+        "Select composer",
+        options=composers,
+        index=3,
+    )
+
+    ids = source_df.composer == selected_composer
+    piece_titles = source_df[ids].title.unique()
+    selected_title = st.selectbox("Select title", options=piece_titles)
+
+    ids = (source_df.composer == selected_composer) & (source_df.title == selected_title)
+    part_df = source_df[ids]
+    part_dataset = midi_dataset.select(part_df.index.values)
+
+    return part_dataset
+
+
+def dataset_configuration():
+    st.header("Dataset Configuration")
+    col1, col2 = st.columns(2)
+    with col1:
+        dataset_path = st.text_input(
+            "Dataset Path",
+            value="roszcz/maestro-sustain-v2",
+            help="Enter the path to the dataset",
+        )
+    with col2:
+        dataset_split = st.selectbox(
+            "Dataset Split",
+            options=["validation", "train", "test"],
+            help="Choose the dataset split to use",
+        )
+
+    dataset = load_hf_dataset(
+        dataset_path=dataset_path,
+        dataset_split=dataset_split,
+    )
+    dataset = select_part_dataset(midi_dataset=dataset)
+
+    st.success(f"Dataset loaded! Total records: {len(dataset)}")
+    return dataset
+
+
+@st.cache_data
+def load_hf_dataset(dataset_path: str, dataset_split: str):
+    dataset = load_dataset(
+        dataset_path,
+        split=dataset_split,
+        trust_remote_code=True,
+        num_proc=8,
+    )
+    return dataset
