+++ requirements.txt
@@ -1,0 +1,9 @@
+datasets>=2.17.1
+fortepyan>=0.2.8
+hydra-core>=1.3.2
+midi_tokenizers @ git+https://github.com/Nospoko/midi-tokenizers
+pandas~=2.1.4
+python-dotenv==1.0.0
+streamlit==1.36.0
+streamlit-pianoroll>=0.7.1
+torch>=2.1.2
